# clash_of_minds_openapi.yaml
openapi: 3.0.3
info:
  title: Clash of Minds – Singularity API
  version: 1.0.0
  description: >
    Bu API, Clash of Minds – Singularity bilgi yarışması oyununun kullanıcı yönetimi,
    oyun oturumu oluşturma, takım ayarlama, yardımcı güç seçimi, kategori seçimi,
    soru yönetimi, puan güncelleme ve sonuç görüntüleme işlemleri için tasarlanmış
    RESTful bir servistir.
  contact:
    name: singularity Team
servers:
  - url: https://api.clashofminds.com
    description: Üretim sunucusu (Production)
  - url: https://staging-api.clashofminds.com
    description: Test sunucusu (Staging)
  - url: http://localhost:3000
    description: Yerel geliştirme sunucusu (Development)
...
security:
  - BearerAuth: []
paths:
...
  /api/game-sessions:
    post:
      tags:
        - Oyun Oturumu
      summary: Oyun Oturumu Oluşturma
      operationId: createGameSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGameSessionRequest'
      responses:
        "201":
          description: Oyun oturumu başarıyla oluşturuldu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameSession'
        "400":
          description: Geçersiz istek verisi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Kimlik doğrulama başarısız
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/game-sessions/{sessionId}/questions/{questionId}:
    parameters:
      - name: sessionId
        in: path
        required: true
        description: Oyun oturumunun benzersiz kimlik numarası
        schema:
          type: string
        example: "sess_001"
      - name: questionId
        in: path
        required: true
        description: Sorunun benzersiz kimlik numarası
        schema:
          type: string
        example: "q_100"
    get:
      tags:
        - Sorular
      summary: Soru Görüntüleme
      operationId: getQuestion
      responses:
        "200":
          description: Soru başarıyla getirildi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
        "404":
          description: Soru veya oyun oturumu bulunamadı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: apiKey
      in: header
      name: Authorization
      description: >
        JWT tabanlı kimlik doğrulama. İstek başlığına
        "Authorization: Bearer <token>" eklenmelidir.
  schemas:
...
    CreateGameSessionRequest:
      type: object
      description: Yeni oyun oturumu oluşturma isteği için gönderilecek veri
      properties:
        gameName:
          type: string
          minLength: 2
          maxLength: 100
          example: "Friday Trivia Night"
        playerCount:
          type: integer
          minimum: 2
          maximum: 12
          example: 4
      required:
        - gameName
        - playerCount
...
    GameSession:
      type: object
      description: Oyun oturumu bilgilerini temsil eden model
      properties:
        id:
          type: string
          example: "sess_001"
        gameName:
          type: string
          example: "Friday Trivia Night"
        playerCount:
          type: integer
          example: 4
        status:
          type: string
          enum: [created, in_progress, finished]
          example: "created"
        activeTeamId:
          type: string
          nullable: true
          example: "team_1"
      required:
        - id
        - gameName
        - playerCount
        - status
...
    Question:
      type: object
      description: Soru modelini temsil eder
      properties:
        id:
          type: string
          example: "q_100"
        categoryId:
          type: string
          example: "cat_history"
        categoryName:
          type: string
          example: "History"
        pointValue:
          type: integer
          example: 200
        text:
          type: string
          example: "Who conquered Constantinople in 1453?"
        correctAnswer:
          type: string
          example: "Mehmed II"
        isDisabled:
          type: boolean
          example: false
      required:
        - id
        - categoryId
        - categoryName
        - pointValue
        - text
        - correctAnswer
        - isDisabled
...
    Error:
      type: object
      description: Hata durumlarında döndürülen standart hata yanıtı
      properties:
        message:
          type: string
          example: "Oyun oturumu bulunamadı"
      required:
        - message